% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip.cat.r
\name{descrip.cat}
\alias{descrip.cat}
\title{Descriptive qualitative explanatory variables}
\usage{
descrip.cat(var)
}
\arguments{
\item{var}{Data vector of qualitative variable}
}
\value{
A data.frame with number(n) and proportion(p) of each category
}
\description{
Calculation of descriptive statistics for qualitative variable.
}
\examples{
x <- rbinom(100,2,0.5)
x <- factor(x, levels=c(0,1,2), labels = c('low','Med','High'))
descrip.cat(x)
}
