% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip.con.r
\name{descrip.con}
\alias{descrip.con}
\title{Descriptive quantitative explanatory variables}
\usage{
descrip.con(var)
}
\arguments{
\item{var}{Data vector of quantitative variable}
}
\value{
A data.frame with n valid, min, max, mean. median, sd and iqr of each variable.
}
\description{
Calculation of descriptive statistics for quantitative variable.
}
\examples{
x <- rnorm(100, mean = 10, sd = 15)
descrip.con(x)
}
