% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorRate.r
\name{ErrorRate}
\alias{ErrorRate}
\title{Error Rate in tables 2x2}
\usage{
ErrorRate(obs, pred)
}
\arguments{
\item{obs}{Data vector of observed values}

\item{pred}{Data vector of predicted values}
}
\value{
A list with accuracy and error rates
}
\description{
Error Rate in tables 2x2 Observed and Predicted values
}
\examples{
x <- rbinom(100,2,0.5)
x <- factor(x, levels=c(0,1,2), labels = c('low','Med','High'))
y <- rbinom(100,2,0.5)
y <- factor(y, levels=c(0,1,2), labels = c('low','Med','High'))
ErrorRate(x,y)
}
