% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsE.r
\name{MsE}
\alias{MsE}
\title{Mean Square Error}
\usage{
MsE(cont.obs, cont.pred)
}
\arguments{
\item{cont.obs}{Data vector of observed quantitative variable}

\item{cont.pred}{Data vector of estimated quantitative variable}
}
\value{
A list with MSE and Error
}
\description{
Calculates the MSE of an estimated quantitative variable over an observed quantitative variable.
}
\examples{
x <- rnorm(100,0,1)
y <- rnorm(100,0,1)
MsE(x,y)
}
