% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.table.style.r
\name{my.table.style}
\alias{my.table.style}
\title{Style table for flextable package}
\usage{
my.table.style(x)
}
\arguments{
\item{x}{flextable object}
}
\value{
format for table
}
\description{
Create pretty tables for MS Word documents
}
\examples{
my.table.style (x)
}
