% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.indicators.r
\name{predictive.indicators}
\alias{predictive.indicators}
\title{Predictive indicators}
\usage{
predictive.indicators(prob, resp, cutoff)
}
\arguments{
\item{prob}{probability of response}

\item{resp}{membership class}

\item{cutoff}{cutoff of probability}
}
\value{
A list of indicators
}
\description{
Function that calculates predictive indicators S, E, Youden, PPV, PNV, LR+ and LR- with their 95% CI.

95% CIs are based on the normal binomial approximation. This approximation is ood if the True Positive

and False Negative are greater than 5, and the patients come from a same sample.
}
\examples{
prob <- sample(seq(0,1,by=0.0001), 200 , replace=T)
resp <- sample(seq(0,1), 200 , replace=T)
cutoff <- seq(0,1,by=0.05)
predictive.indicators(prob,resp,cutoff)
}
