% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip.cat.r, R/descrip.con.r
\docType{package}
\name{Quesadilla-package}
\alias{Quesadilla-package}
\title{Quesadilla package}
\description{
This package contains some functions applied in Biostatistics

This package contains some functions applied in Biostatistics
}
