% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCarea.r
\name{ROCarea}
\alias{ROCarea}
\title{AUC area}
\usage{
ROCarea(bin.var, prob.var)
}
\arguments{
\item{bin.var}{binary classification variable}

\item{prob.var}{quantitative variable that can be a probability, total questionnaire score,..}
}
\value{
A list with AUC and 95% CI
}
\description{
Function that calculates the area under the ROC curve, from a binary classification variable and a quantitative variable
}
\examples{
prob <- sample(seq(0,1,by=0.0001), 200 , replace=T)
resp <- sample(seq(0,1), 200 , replace=T)
ROCarea(resp,prob)
}
