% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.con.plus.r
\name{test.con.plus}
\alias{test.con.plus}
\title{Anova and Kruskall-Wallis test}
\usage{
test.con.plus(x, evento, ncat)
}
\arguments{
\item{x}{Explanatory quantitative variable}

\item{evento}{Response qualitative variable}

\item{ncat}{number of categories of response variable (ncat > 2)}
}
\value{
A data.frame with number, mean and SD in ncat populations, p-value and method Anova or Kruskall-Wallis test
}
\description{
Calculation mean and SD in 3 populations, with Anova or Kruskall-Wallis test
}
\examples{
x <- rnorm(100,0,1)
y <- rbinom(100,2,0.5)
y <- factor(y, levels=c(0,1,2), labels = c('low','Med','High'))
test.con.plus(x,y,3)
}
