% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.con.r
\name{test.con}
\alias{test.con}
\title{t-test and Wilcoxon test}
\usage{
test.con(x, evento)
}
\arguments{
\item{x}{Explanatory quantitative variable}

\item{evento}{Response qualitative variable with 2 categories}
}
\value{
A data.frame with number, mean and SD in 2 populations, p-value and method t-test or Wilcoxon test
}
\description{
Calculation mean and SD in 2 populations, with t-test or Wilcoxon test
}
\examples{
x <- rnorm(100,0,1)
y <- rbinom(100,1,0.5)
y <- factor(y, levels=c(0,1), labels = c('low','High'))
test.con(x,y)
}
